Preserve set -e in bin/setup
Command cp -nv exits with 1 if the file already exists, which is something
that could easily occur if setup is called on an already initialized
repository or files have been copied before for reasons.

This behavior was not compatible with set -e because it would stop the
execution of the setup, so it was not kept.
However, adding "|| true" at the end of the cp lines ensures those
commands always return 0, hence allowing the use of set -e, which is
a best practice of setup scripts.